Add tests to util/div.c

+++ include/div.h
@@ -1,8 +1,6 @@
 #ifndef _DIV_H_
 #define _DIV_H_
 
-#include <stdlib.h>
-
-size_t round_up_div(size_t a, size_t b);
+int round_up_div(int a, int b);
 
 #endif
